+++ src/shared/db/databaseUtil.ts
@@ -7,7 +7,7 @@ import {
   OVERLAY_SEEN,
   OVERLAY_DRAFT,
   OVERLAY_LOG
-} from "./constants";
+} from "../constants";
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 // schema definitions and default values
@@ -202,40 +202,4 @@ export function defaultCallback(
   } else {
     console.log(`Local database: document ${verb}`);
   }
-}
-
-export class DatabaseNotInitializedError extends Error {
-  constructor() {
-    super("LocalDatabase has not been initialized.");
-    this.name = "DatabaseNotInitializedError";
-  }
-}
-
-export interface LocalDatabase {
-  dbName: string;
-  filePath: string;
-  find(
-    table: string,
-    key: string,
-    callback?: (err: Error | null, data: any) => void
-  ): void;
-  findAll(callback?: (err: Error | null, data: any) => void): void;
-  init(dbName: string, arenaName?: string): void;
-  upsert(
-    table: string,
-    key: string,
-    data: any,
-    callback?: (err: Error | null, num: number) => void,
-    globals?: any
-  ): void;
-  upsertAll(
-    data: any,
-    callback?: (err: Error | null, num: number) => void,
-    intermediateCallback?: (err: Error | null, num: number) => void
-  ): void;
-  remove(
-    table: string,
-    key: string,
-    callback?: (err: Error | null, num: number) => void
-  ): void;
 }
