+++ src/window_background/background.js
@@ -4,8 +4,8 @@ import path from "path";
 import fs from "fs";
 import sha1 from "js-sha1";
 import * as httpApi from "./http-api";
-import { appDb, playerDb } from "../shared/LocalDatabase";
-import { rememberDefaults } from "../shared/databaseUtil";
+import { appDb, playerDb } from "../shared/db/LocalDatabase";
+import { rememberDefaults } from "../shared/db/databaseUtil";
 import playerData from "../shared/player-data";
 import { getReadableFormat } from "../shared/util";
 import { HIDDEN_PW, MAIN_DECKS } from "../shared/constants";
